Numeric and short anwser question types work

+++ MoodleSyntaxGenerator/MoodleSyntaxGenerator/Controllers/Controller.cs
@@ -15,7 +15,7 @@ namespace MoodleSyntaxGenerator.Controllers
 			_generator = generator;
 		}
 
-		public string GenerateNumeric(string question, int answer, decimal tolerance = 0)
+		public string GenerateNumeric(string question, double answer, decimal tolerance = 0)
 		{
 			return _generator.GenerateNumeric(question, answer, tolerance);
 		}
